Fix schema names, create protocol enum
The case of the schema names was not working with the generated
TypeScript code - imports were using the wrong case and were therefore
broken. This fixes that issue.

Also, the `Check` type is an enum of each protocol, but it cannot be
used easily in templates. An additional `Protocol` enum was added to
display the check type.